# Cache transformed keys