I noticed Alba was quite slow with the keys transformation feature enabled. After a quick review of the code, I decided to add a memoization mechanism directly in the main `Alba` module. This makes the code more transparent. While putting the memoization at the resource-level cache was also an option, it would still require checking `Alba.inflector` and `Alba.symbolize_keys`, which would look less obvious.

Additionally, I've updated the benchmark: it now supports Ruby 3.4, Rails and Representable serializers pass the output checks. I've updated the README (without the new results - I think it's better for @okuramasafumi to run them, especially interesting to see the latest `json` gem being faster than `Oj.optimize_rails` 😄). Finally, the script now displays the gem versions used at the end.

Benchmark results (we can drop transformation from benchmark once this PR is reviewed).

Note that `alba_with_transformation` is faster since the "pure" version calls `regularize_key` all the time.

### BEFORE

#### ips:

```
Comparison:
               panko:      585.8 i/s
                alba:      186.5 i/s - 3.14x  slower
alba_with_transformation:       84.2 i/s - 6.96x  slower
```

#### memory:

```
Comparison:
               panko:     259178 allocated
                alba:    1002313 allocated - 3.87x more
alba_with_transformation:    2331053 allocated - 8.99x more
```


### AFTER

#### ips:

```

Comparison:
               panko:      598.3 i/s
alba_with_transformation:      212.1 i/s - 2.82x  slower
                alba:      186.3 i/s - 3.21x  slower
```

#### memory:

```
Comparison:
               panko:     259178 allocated
alba_with_transformation:     826253 allocated - 3.19x more
                alba:    1002313 allocated - 3.87x more
```