@okuramasafumi 2024-12-27
Hi @skryukov, thank you for the patch! I didn't notice transforming keys is one of the bottlenecks. I highly appreciate it.
I think `Alba.inflector` rarely changes during runtime, so we might refactor it a bit:

```diff
diff --git a/lib/alba.rb b/lib/alba.rb
index afcf8de..e3626f2 100644
--- a/lib/alba.rb
+++ b/lib/alba.rb
@@ -117,6 +117,7 @@ module Alba
     #   When it's a Class or a Module, it should have some methods, see {Alba::DefaultInflector}
     def inflector=(inflector)
       @inflector = inflector_from(inflector)
+      reset_transform_keys
     end
 
     # @param block [Block] resource body
@@ -145,6 +146,7 @@ module Alba
     # Configure Alba to symbolize keys
     def symbolize_keys!
       @symbolize_keys = true
+      @_transformed_keys = nil
     end
 
     # Configure Alba to stringify (not symbolize) keys
@@ -171,21 +173,19 @@ module Alba
     def transform_key(key, transform_type:)
       raise Alba::Error, 'Inflector is nil. You must set inflector before transforming keys.' unless inflector
 
-      @_transformed_keys ||= Hash.new { |h, k| h[k] = Hash.new { |h2, k2| h2[k2] = Hash.new { |h3, k3| h3[k3] = {} } } }
-      @_transformed_keys[inflector][transform_type][@symbolize_keys][key] ||=
-        begin
-          key = key.to_s
-
-          k = case transform_type
-              when :camel then inflector.camelize(key)
-              when :lower_camel then inflector.camelize_lower(key)
-              when :dash then inflector.dasherize(key)
-              when :snake then inflector.underscore(key)
-              else raise Alba::Error, "Unknown transform type: #{transform_type}"
-              end
-
-          regularize_key(k)
-        end
+      @_transformed_keys[transform_type][key] ||= begin
+        key = key.to_s
+
+        k = case transform_type
+            when :camel then inflector.camelize(key)
+            when :lower_camel then inflector.camelize_lower(key)
+            when :dash then inflector.dasherize(key)
+            when :snake then inflector.underscore(key)
+            else raise Alba::Error, "Unknown transform type: #{transform_type}"
+            end
+
+        regularize_key(k)
+      end
     end
 
     # Register types, used for both builtin and custom types
@@ -213,7 +213,7 @@ module Alba
       @_on_error = :raise
       @_on_nil = nil
       @types = {}
-      @_transformed_keys = nil
+      reset_transform_keys
       register_default_types
     end
 
@@ -320,6 +320,10 @@ module Alba
         register_type(:"ArrayOf#{t}", check: ->(d) { d.is_a?(Array) && d.all? { _1.is_a?(t) } })
       end
     end
+
+    def reset_transform_keys
+      @_transformed_keys = Hash.new { |h, k| h[k] = {} }
+    end
   end
 
   reset!
```

This reduces nesting of a cache hash and instead resets the cache after some methods (`Alba.inflector=`, for example).
In my local environment (thank you for the benchmark improvement, too!), this refactoring makes it even faster than the original (maybe I'm wrong).
What do you think?