@okuramasafumi 2024-12-27
I noticed an interesting fact that with YJIT enabled, `alba_with_transformation` is slower than `alba`. Of course in general they are much faster than without YJIT, but I don't know how JIT changes the benchmark result in this situation.