@skryukov 2024-12-28
Thanks for the review @okuramasafumi, I applied your patch (also called `reset_transform_keys` on both `symbolize_keys!` and `stringify_keys!`).

For me, running on Ruby 3.4.1 `alba_with_transformation` is still faster then `alba` 😅

YJIT + OJ:
```
Comparison:
               panko:      842.8 i/s
       turbostreamer:      429.9 i/s - same-ish: difference falls within error
         jserializer:      363.6 i/s - 2.32x  slower
alba_with_transformation:      347.6 i/s - 2.43x  slower
                alba:      300.4 i/s - 2.81x  slower
     fast_serializer:      244.7 i/s - 3.44x  slower
               rails:      168.6 i/s - 5.00x  slower
         blueprinter:      154.1 i/s - 5.47x  slower
       representable:      101.4 i/s - 8.31x  slower
          simple_ams:       94.5 i/s - 8.92x  slower
                 ams:       27.5 i/s - 30.70x  slower
         alba_inline:       14.5 i/s - 58.04x  slower
```

YJIT without Oj:
```
Comparison:
               panko:      857.3 i/s
       turbostreamer:      422.9 i/s - 2.03x  slower
alba_with_transformation:      413.0 i/s - 2.08x  slower
                alba:      379.8 i/s - 2.26x  slower
         jserializer:      370.2 i/s - 2.32x  slower
     fast_serializer:      244.4 i/s - 3.51x  slower
               rails:      161.3 i/s - 5.31x  slower
         blueprinter:      146.4 i/s - 5.86x  slower
       representable:      100.7 i/s - 8.52x  slower
          simple_ams:       87.9 i/s - 9.75x  slower
                 ams:       25.2 i/s - 33.96x  slower
         alba_inline:       14.4 i/s - 59.61x  slower
```