@okuramasafumi 2024-12-28
@skryukov Thank you, I cannot see your new commit on GitHub, maybe you didn't push it?
And for benchmark result, it's quite interesting. Alba might be faster than Oj? If so, I'd like to add this to README :)