@skryukov 2024-12-28
Ooops, just pushed, sorry 😂 

Regarding Oj: Yes, it now makes sense to compare JSON >= 2.9 with Oj. However, I believe the results will depend on the data. My colleague showed me some benchmarks where flat responses with simpler data types might still perform better when using Oj.